feat(pixi.toml): add all `package.xml` depends
WARNING: THIS DROPS WINDOWS SUPPORT!

many dependencies (in particular, the Gazebo simulator, `rqt`, and
laserscan stuff) don't seem to be built for Windows.

also, do you see that giant `ros-humble-gazebo-ros2-control` TODO in
`pixi.toml`? we need to fix that. it was annoying on the actual Rover,
but now it's just ridiculous